holy crap bruh

+++ throttle/src/sequencer.h
@@ -3,6 +3,6 @@
 
 #include <vector>
 
-int sequencer_start_trace(int sock, int gap, std::vector<double> bps);
+int sequencer_start_trace(int sock, int gap, std::vector<float> bps);
 
 #endif //CLOVER_SEQUENCER_H
